@@ -1,15 +1,12 @@
 package com.baeldung.repository;
 
-import com.baeldung.model.BasicUser;
-import org.springframework.data.jpa.repository.EntityGraph;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<BasicUser, Long> {
+import org.springframework.data.jpa.repository.JpaRepository;
 
-    @EntityGraph(attributePaths = "permissions")
-    Optional<BasicUser> findDetailedByUsername(String username);
+import com.baeldung.model.BasicUser;
+
+public interface UserRepository extends JpaRepository<BasicUser, Long> {
 
     Optional<BasicUser> findSummaryByUsername(String username);
 
