@@ -1,11 +1,9 @@
 package com.baeldung.web.controller;
 
 import com.baeldung.web.dto.Foo;
-import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -13,22 +11,15 @@ import javax.servlet.http.HttpServletResponse;
 import static org.apache.commons.lang3.RandomStringUtils.randomAlphabetic;
 import static org.apache.commons.lang3.RandomStringUtils.randomNumeric;
 
-@Controller
+@RestController
 public class FooController {
 
-    public FooController() {
-        super();
-    }
-
-    // API - read
-    @RequestMapping(method = RequestMethod.GET, value = "/foos/{id}")
-    @ResponseBody
+    @GetMapping("/foos/{id}")
     public Foo findById(@PathVariable final long id, HttpServletRequest req, HttpServletResponse res) {
-        // System.out.println(req.getHeaderNames());
-        // System.out.println("------" + req.getHeader("Test"));
         if (req.getHeader("Test") != null) {
             res.addHeader("Test", req.getHeader("Test"));
         }
+
         return new Foo(Long.parseLong(randomNumeric(2)), randomAlphabetic(4));
     }
 
