@@ -1,21 +1,14 @@
 package com.baeldung.web.dto;
 
 public class Foo {
+
     private long id;
     private String name;
 
-    public Foo() {
-        super();
-    }
-
     public Foo(final long id, final String name) {
-        super();
-
         this.id = id;
         this.name = name;
     }
-
-    //
 
     public long getId() {
         return id;
