@@ -1,12 +1,10 @@
 package com.baeldung.controller.optionalpathvars;
 
-import static com.baeldung.model.Article.DEFAULT_ARTICLE;
+import static com.baeldung.controller.optionalpathvars.Article.DEFAULT_ARTICLE;
 
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.baeldung.model.Article;
 
 @RestController
 @RequestMapping(value = "/seperateMethods")
