@@ -1,6 +1,7 @@
 package com.baeldung.rest.karate;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
+import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import com.intuit.karate.junit4.Karate;
 import cucumber.api.CucumberOptions;
 import org.junit.AfterClass;
@@ -14,13 +15,15 @@ import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 @CucumberOptions(features = "classpath:karate")
 public class KarateIntegrationTest {
 
-    private static final WireMockServer wireMockServer = new WireMockServer();
+    private static final int PORT_NUMBER = 8097;
+
+    private static final WireMockServer wireMockServer = new WireMockServer(WireMockConfiguration.options().port(PORT_NUMBER));
 
     @BeforeClass
     public static void setUp() throws Exception {
         wireMockServer.start();
 
-        configureFor("localhost", 8080);
+        configureFor("localhost", PORT_NUMBER);
         stubFor(get(urlEqualTo("/user/get"))
                 .willReturn(aResponse()
                         .withStatus(200)
