@@ -1,12 +1,14 @@
 package com.baeldung.calculator;
 
-import com.baeldung.cucumber.Calculator;
-import cucumber.api.java.Before;
-import cucumber.api.java.en.Given;
-import cucumber.api.java.en.Then;
-import cucumber.api.java.en.When;
 import org.hamcrest.Matchers;
 import org.junit.Assert;
+
+import com.baeldung.cucumber.Calculator;
+
+import io.cucumber.java.Before;
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
 
 public class CalculatorRunSteps {
 
